Potential fix for code scanning alert no. 77: Multiplication result converted to larger type

+++ hw/net/e1000e_core.c
@@ -935,8 +935,9 @@ e1000e_has_rxbufs(E1000ECore *core, const E1000ERingInfo *r,
     trace_e1000e_rx_has_buffers(r->idx, bufs, total_size,
                                 core->rx_desc_buf_size);
 
-    return total_size <= bufs / (core->rx_desc_len / E1000_MIN_RX_DESC_LEN) *
-                         core->rx_desc_buf_size;
+    return total_size <= ((size_t)bufs /
+                          (core->rx_desc_len / E1000_MIN_RX_DESC_LEN)) *
+                         (size_t)core->rx_desc_buf_size;
 }
 
 void
